+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### 🚀 Added
+- Add check: Unordered keys [#72](https://github.com/mgrachev/dotenv-linter/pull/72)
 - Add the ability to specify the directory where to run [#65](https://github.com/mgrachev/dotenv-linter/pull/65) ([@Louis-Lesage](https://github.com/Louis-Lesage))
 - Add check: Duplicated keys [#33](https://github.com/mgrachev/dotenv-linter/pull/33) ([@mstruebing](https://github.com/mstruebing))
 - Add exit with the code 1 on warnings found [#58](https://github.com/mgrachev/dotenv-linter/pull/58) ([@Louis-Lesage](https://github.com/Louis-Lesage))