+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add skipping of commented or empty lines [#37](https://github.com/mgrachev/dotenv-linter/pull/37) ([@mstruebing](https://github.com/mstruebing))
 
 ### 🔧 Changed
+- Rename `leading_space` to `leading_character` and check for allowed chars [#63](https://github.com/mgrachev/dotenv-linter/pull/63) ([@mstruebing](https://github.com/mstruebing))
 - Remove multiple checks of the same file [#62](https://github.com/mgrachev/dotenv-linter/pull/62) ([@mstruebing](https://github.com/mstruebing))
 - Add mutability support for checks [#52](https://github.com/mgrachev/dotenv-linter/pull/52)